Move remaining non-view functions to domain module
